+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 ### Version 6.5
 - Remove screenshots directory
+- Add show/hide password in user profile
 
 ### Version 6.4
 - Fix supplier details edit modal